+++ ziplime/config/base_algorithm_config.py
@@ -1,8 +1,4 @@
-import datetime
-from functools import lru_cache
-
-from exchange_calendars import get_calendar, ExchangeCalendar
-from pydantic import BaseModel, validator, field_validator
+from pydantic import BaseModel
 
 
 class BaseAlgorithmConfig(BaseModel):